Add support for WORDCHARS when spell checking a string
Hunspell affix files have a WORDCHARS property which is used for specifying characters that can be part of a word in that language. This setting must be taken into account when looking for words or words may be marked as misspelled when in fact they aren't. Some examples:

FR: -’'1234567890.
SV: -0123456789.:
EN: 0123456789

For example, in Swedish `t.ex.` is a word which is why `.` is included in WORDCHARS. `t.ex` will be marked as misspelled unless WORDCHARS is honored.

Since the Hunspell property for accessing the WORDCHARS value is not exposed to the C API it cannot be access from the LCB file. The `hunspellLoadWordCharsFromAffixFile` command has been added which will search an affix file for WORDCHARS and load the value for use in `hunspellFindMisspelledWords`.